Enable custom log out functionality
Reverse proxies often include ways to log out of the portal that you are
logged in to. Some organizations have security requirements that require
the ability to log out from any application. This feature adds the ability
to log out and redirect to another page, presumably to log out of a
portal.