Require Jenkins 2.479.1 or newer
Does not migrate from acegi security to Spring Security.  Hopes that an
acegi Authorization object can be cast to a Spring Authorization object.
That seems highly unlikely, but the alternative is to migrate from acegi
security to Spring Security and that is much more work.

Passes automated tests with Java 21.  Not tested in any interactive tests.
Likely will fail real use in many horrible ways.